Bring in more recent changes from UPP
Brings in additional changes in the scripts we are bringing over from
UPP.

Refs: #766

+++ scripts/exgfs_atmos_nceppost.sh
@@ -25,6 +25,12 @@ echo " Feb 18 - Meng - Removed legacy setting for generating grib1 data"
 echo "                 and reading sigio model outputs."
 echo " Aug 20 - Meng - Remove .ecf extentsion per EE2 review."
 echo " Sep 20 - Meng - Update clean up files per EE2 review."
+echo " Dec 20 - Meng - Add alert for special data file."
+echo " Mar 21 - Meng - Update POSTGRB2TBL default setting."
+echo " Jun 21 - Mao  - Instead of err_chk, catch err and print out"
+echo "                 WAFS failure warnings to avoid job crashing"
+echo " Oct 21 - Meng - Remove jlogfile for wcoss2 transition."
+echo " Feb 22 - Lin - Exception handling if anl input not found."
 echo "-----------------------------------------------------"
 #####################################################################
 
@@ -34,7 +40,7 @@ cd $DATA
 
 # specify model output format type: 4 for nemsio, 3 for sigio
 msg="HAS BEGUN on $(hostname)"
-postmsg "$jlogfile" "$msg"
+postmsg "$msg"
 
 export POSTGPSH=${POSTGPSH:-$USHgfs/gfs_nceppost.sh}
 export GFSDOWNSH=${GFSDOWNSH:-$USHgfs/fv3gfs_downstream_nems.sh}
@@ -203,31 +209,38 @@ then
       export PGIOUT=wafsifile
 
       $POSTGPSH
-      export err=$?; err_chk
+      export err=$?
+      if [ $err -ne 0 ] ; then
+        echo " *** GFS POST WARNING: WAFS output failed for analysis, err=$err"
+      else
+        # WAFS package doesn't process this part.
+        # Need to be saved for WAFS U/V/T verification, 
+        # resolution higher than WAFS 1.25 deg for future compatibility
+        wafsgrid="latlon 0:1440:0.25 90:721:-0.25"
+        $WGRIB2 $PGBOUT -set_grib_type same -new_grid_winds earth \
+                -new_grid_interpolation bilinear -set_bitmap 1 \
+  	      -new_grid $wafsgrid ${PGBOUT}.tmp
 
-      # WAFS package doesn't process this part.
-      # Need to be saved for WAFS U/V/T verification, 
-      # resolution higher than WAFS 1.25 deg for future compatibility
-      wafsgrid="latlon 0:1440:0.25 90:721:-0.25"
-      $WGRIB2 $PGBOUT -set_grib_type same -new_grid_winds earth \
-              -new_grid_interpolation bilinear -set_bitmap 1 \
-	      -new_grid $wafsgrid ${PGBOUT}.tmp
+        if test $SENDCOM = "YES"
+        then
+           cp ${PGBOUT}.tmp $COMOUT/${PREFIX}wafs.0p25.anl
+           $WGRIB2 -s ${PGBOUT}.tmp > $COMOUT/${PREFIX}wafs.0p25.anl.idx
 
-      if test $SENDCOM = "YES"
-      then
-         cp ${PGBOUT}.tmp $COMOUT/${PREFIX}wafs.0p25.anl
-         $WGRIB2 -s ${PGBOUT}.tmp > $COMOUT/${PREFIX}wafs.0p25.anl.idx
-
-         if [ $SENDDBN = YES ]; then
-            $DBNROOT/bin/dbn_alert MODEL GFS_WAFS_GB2 $job $COMOUT/${PREFIX}wafs.0p25.anl
-            $DBNROOT/bin/dbn_alert MODEL GFS_WAFS_GB2__WIDX $job $COMOUT/${PREFIX}wafs.0p25.anl.idx
-         fi
+           # if [ $SENDDBN = YES ]; then
+           #    $DBNROOT/bin/dbn_alert MODEL GFS_WAFS_GB2 $job $COMOUT/${PREFIX}wafs.0p25.anl
+           #    $DBNROOT/bin/dbn_alert MODEL GFS_WAFS_GB2__WIDX $job $COMOUT/${PREFIX}wafs.0p25.anl.idx
+           # fi
+        fi
+        rm $PGBOUT ${PGBOUT}.tmp
       fi
-      rm $PGBOUT ${PGBOUT}.tmp
    fi
   fi
 ##########################  WAFS U/V/T analysis end  ##########################
-
+else
+  #### atmanl file not found need failing job
+  echo " *** FATAL ERROR: No model anl file output "
+  export err=9
+  err_chk
 fi
 
 #----------------------------------
@@ -275,7 +288,7 @@ do
     set -x
 
     msg="Starting post for fhr=$fhr"
-    postmsg "$jlogfile" "$msg"
+    postmsg "$msg"
 
     ###############################
     # Put restart files into /nwges 
@@ -301,7 +314,7 @@ do
     export GFSOUT=${PREFIX}gfsio${fhr}
 
     if [ $GRIBVERSION = 'grib2' ]; then
-      export POSTGRB2TBL=${POSTGRB2TBL:-${G2TMPL_SRC}/params_grib2_tbl_new}
+      export POSTGRB2TBL=${POSTGRB2TBL:-${g2tmpl_ROOT}/share/params_grib2_tbl_new}
       export PostFlatFile=${PostFlatFile:-$PARMpost/postxconfig-NT-GFS.txt}
 
       if [ $RUN = gfs ] ; then
@@ -432,7 +445,7 @@ do
       then                    
         export FH=$(expr $fhr + 0)     
         $GFSDOWNSHF                  
-        export err=$?; err_chk      
+        export err=$?; err_chk
       fi   
 
       if [ $INLINE_POST = ".false." ]; then
@@ -498,6 +511,10 @@ do
        mv goesfile $COMOUT/${SPECIALFL}f$fhr
        mv goesifile $COMOUT/${SPECIALFLIDX}f$fhr
 
+       if [ $SENDDBN = YES ]; then
+           $DBNROOT/bin/dbn_alert MODEL GFS_SPECIAL_GB2 $job $COMOUT/${SPECIALFL}f$fhr
+       fi
+
     fi
     fi
 # end of satellite processing
@@ -535,14 +552,17 @@ do
 	  fi
 
           export err=$?; err_chk
-
-          if [ -e $PGBOUT ]
-          then
-          if test $SENDCOM = "YES"
-          then
-              cp $PGBOUT $COMOUT/${PREFIX}wafs.grb2f$fhr
-              cp $PGIOUT $COMOUT/${PREFIX}wafs.grb2if$fhr
-          fi
+          if [ $err -ne 0 ] ; then
+                    echo " *** GFS POST WARNING: WAFS output failed for f${fhr}, err=$err"
+          else
+            if [ -e $PGBOUT ]
+            then
+              if test $SENDCOM = "YES"
+              then
+                  cp $PGBOUT $COMOUT/${PREFIX}wafs.grb2f$fhr
+                  cp $PGIOUT $COMOUT/${PREFIX}wafs.grb2if$fhr
+              fi
+            fi
           fi
       fi
       [[ -f wafsfile ]] && rm wafsfile ; [[ -f wafsifile ]] && rm wafsifile
